Crate locha_p2p

Locha P2P

This library contains the necessary code to set up a node that works with the Locha P2P Chat.

Modules

identity
Identity

msg

p2p

runtime
Runtime

upnp
UPnP

Structs

Multiaddr
Representation of a Multiaddr.

PeerId
Identifier of a peer of the network.

Enums

Protocol
Protocol describes all possible multiaddress protocols.

Functions

build_transport
Builds the Transport used in Locha P2P