Module locha_p2p::runtime

Runtime

This contains a runtime for the Chat. The network behaviour and swarm is handled on a separate thread, we communicate with the thread by sending actions to it. It also reports us throhgh callbacks the events that happen. This is especially useful for language bridges such as the JNI or Node.JS where we can't use the async/await method as we do on Rust and we need a thread to do it for us, that's what Runtime does.

Examples

use locha_p2p::identity::Identity;
use locha_p2p::runtime::events::RuntimeEvents;
use locha_p2p::runtime::config::RuntimeConfig;
use locha_p2p::runtime::Runtime;
use locha_p2p::{Multiaddr, PeerId};

struct EventsHandler;

impl RuntimeEvents for EventsHandler {
    fn on_new_message(&mut self, peer_id: &PeerId, message: Vec<u8>) {
        println!("new message: {:?}", message);
    }
}

let identity = Identity::generate();

let config = RuntimeConfig {
    identity,
    listen_addr: "/ip4/0.0.0.0/tcp/0".parse().expect("invalid address"),
    channel_cap: 20,
    heartbeat_interval: 5,

    upnp: false,
    mdns: false,

    bootstrap_nodes: Vec::new(),
};

let (runtime, runtime_task) = Runtime::new(config, Box::new(EventsHandler)).unwrap();

async_std::task::spawn(runtime_task);

Modules

config
Runtime configuration

error
Runtime errors

events
Runtime events

Structs

Addresses
A non-empty list of (unique) addresses of a peer in the routing table.

EntryView
A cloned, immutable view of an entry that is either present in a bucket or pending insertion.

Key
A Key in the DHT keyspace with preserved preimage.

NetworkInfo
Information about the network obtained by [Network::info()].

Runtime
Locha P2P runtime

Type Definitions

Kbucket
An owned Kademlia K-bucket