Module locha_p2p::upnp

UPnP

This is the Universal Plug and Play (UPnP) NetworkBehaviour, this behaviour reports our external IPv4 address if we are behind a NATed router or a general IGD. It also maps all of the open TCP ports for the IPv4 listening addresses found.

Each 20 minutes the address is checked and if it differs it's reported again. Also the ports get remapped in case a buggy UPnP implementation drops the mapping, other application messes with it, etc.

Usage

use locha_p2p::upnp::UpnpBehaviour;

const PROTOCOL_DESCRIPTION: &str = "My Great Protocol";

let _upnp = UpnpBehaviour::new(PROTOCOL_DESCRIPTION.into());

Structs

UpnpBehaviour